backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
aiofiles==23.2.1
openai>=1.59.0
python-dotenv